requirements.txt
pyhumps
fastapi
uvicorn
pytest
fastapi-sqlalchemy
pydantic
psycopg2